test delete reminder
