+++ Annoy-o-Bot.AcceptanceTests/Fakes/FakeGithubInstallation.cs
@@ -1,7 +1,7 @@
 ﻿using Annoy_o_Bot.GitHub;
 using Octokit;
 
-namespace Annoy_o_Bot.AcceptanceTests;
+namespace Annoy_o_Bot.AcceptanceTests.Fakes;
 
 class FakeGithubInstallation : IGitHubAppInstallation
 {
@@ -13,21 +13,21 @@ class FakeGithubInstallation : IGitHubAppInstallation
 
     public Task Initialize(long installationId)
     {
-        this.InstallationId = installationId;
-        this.Initialized = true;
+        InstallationId = installationId;
+        Initialized = true;
         return Task.CompletedTask;
     }
 
     private readonly Dictionary<string, string> files = new();
     public void AddFileContent(string filePath, string content)
     {
-        this.files.Add(filePath, content);
+        files.Add(filePath, content);
     }
 
     //TODO test behavior when file does not exist
     public Task<string> ReadFileContent(string filePath, long repositoryId, string branchReference)
     {
-        return Task.FromResult(this.files[filePath]);
+        return Task.FromResult(files[filePath]);
     }
 
     public Task CreateCheckRun(NewCheckRun checkRun, long repositoryId)
