+++ Annoy-o-Bot/CosmosDB/ICosmosClientWrapper.cs
@@ -6,6 +6,6 @@ namespace Annoy_o_Bot.CosmosDB;
 
 public interface ICosmosClientWrapper
 {
-    Task<ReminderDocument> LoadReminder(IDocumentClient cosmosClient, string fileName, long installationId,
-        long repositoryId);
+    Task<ReminderDocument?> LoadReminder(IDocumentClient cosmosClient, string fileName, long installationId, long repositoryId);
+    Task Delete(IDocumentClient cosmosClient, string fileName, long installationId, long repositoryId);
 }
